Mapper 159 - fix address increment to work "correctly"
The chip is actually big endian, but Bandai treated it as little endian;
to retain save compatibility, I'm leaving it little-endian while fixing
multi-byte read/write to increment the actual big-endian address (which
requires reversing the bits before and after).

If I had included a save version indicator in the mapper savestate logic,
I could've fixed the save data on load and avoided this mess.