Change NES-DxROM to use the shared N118 module
This required adding a separate SaveLoad routine to handle the fact that
DxROM's local N108 module was storing pre-shifted bank numbers for CHR
$0000-$0FFF while the shared N118 module was storing masked values and
then shifting them during Sync.

This also (properly) restricts the NES-DxROM boards to 128KB of PRG ROM.

This is also the first usage of a non-zero mapper data version.